// src/lib/supabase/types.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[];

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1";
  };
  public: {
    Tables: {
// ...
      contacts: {
        Row: {
          created_at: string;
          id: string;
          message: string;
          metadata: Json | null;
          name: string;
          phone: string;
          status: string | null;
          user_id: string;
        };
        Insert: {
          created_at?: string;
          id?: string;
          message: string;
          metadata?: Json | null;
          name: string;
          phone: string;
          status?: string | null;
          user_id: string;
        };
        Update: {
          created_at?: string;
          id?: string;
          message?: string;
          metadata?: Json | null;
          name?: string;
          phone?: string;
          status?: string | null;
          user_id?: string;
        };
        Relationships: [
          {
            foreignKeyName: "contacts_user_id_fkey";
            columns: ["user_id"];
            isOneToOne: false;
            referencedRelation: "profiles";
            referencedColumns: ["id"];
          },
        ];
      };
// ...
    };
    Views: {
      [_ in never]: never;
    };
    Functions: {
      increment_campaign_sent: { Args: { row_id: string }; Returns: undefined };
    };
    Enums: {
      whatsapp_connection_type: "webhook" | "evolution";
    };
    CompositeTypes: {
      [_ in never]: never;
    };
  };
};